requirements.txt
# Requirements for Poochable 
Django == 1.5
djangorestframework == 2.2.6
django-storages == 1.1.8
easy-thumbnails == 1.2
